+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 
 # Version 0.0.2 (wip)
+- Fixed popup links
+- Fleet list quick filter by sector
+- Bugfix fleet list quick filter
 - Copy to cliboard in fleet page
 - Bugfix surface scan